Correction type

+++ src/main/java/cat/itacademy/barcelonactiva/Gispert/Judith/s04/t02/n02/Repository/IFruitRepo.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface IFruitRepo extends JpaRepository<Fruit, Float> { }
+public interface IFruitRepo extends JpaRepository<Fruit, Integer> { }
